Start using resource dispatch and define Article.
The first few steps relied on the defualt dispatcher, object dispatch,
to look up the "endpoint" to call. We're beginning to get real structure
in here, so we've switched to resource dispatch to better model a REST
interface.

This required the `__call__` method on the `Wiki` root object to be
re-named `get` and the addition of some additional initializer
arguments.  The `__getitem__` method added is a stub, only pretending to
load data from a database, and the stub `Article.get` method just lets
us know the name stored in that stub, for now.